Move status enums to `reports` component.
