+++ app/schemas/__init__.py
@@ -1,2 +1,1 @@
-from .enums import *
 from .roles import UserRole
